1. Добавил генерацию для партиций 2. Выполнил мелки корректировки 3. Внедрил и настроил партицию, перенастроил роли 4. Добавил форму авторизации 5. Внедрил JSON 6.  Исправил поиск по колонкам

+++ CourseWork PSwDBS Pankov/OperationPages/RequestPages/request_transportation_Page.xaml.cs
@@ -42,7 +42,7 @@ namespace CourseWork_PSwDBS_Pankov.OperationPages.RequestPages
             {
                 var costObj = dbContext.ExecuteScalar<object>($"SELECT * FROM getTransportationCostByCargoWeight('{WeightTextBox.Text}')");
 
-                if(costObj is int)
+                if(costObj is long)
                 {
                     CostTranspoerationLable.Content = "Общая стоимость: " + costObj;
                 }
